refactor: methods response sending, route package
