+++ cmd/server/handler/routes/routes.go
@@ -1,12 +1,16 @@
-package handler
+package routes
 
 import (
 	"net/http"
+
+	"github.com/doruo/falloutdle/cmd/server/handler"
 )
 
 func SetupRoutes(mux *http.ServeMux) {
-	handler := NewGameHandler()
+	handler := handler.NewGameHandler()
+
 	mux.HandleFunc("/", handler.HandleGetHome)
 	mux.HandleFunc("/api/today", handler.HandleGetTodayCharacter)
 	mux.HandleFunc("/api/random", handler.HandleGetRandomCharacter)
+	mux.HandleFunc("/api/guess", handler.HandlePostGuessCharacter)
 }
